esercizi component elenco da server
