+++ vitoswimpt.client/src/app/app.component.spec.ts
@@ -29,17 +29,7 @@ describe('AppComponent', () => {
   });
 
   it('should retrieve training data from the server', () => {
-    const mockEsercizi = [
-      { ripetizioni: '4', distanza: 200, recupero: 60, stile: 'dorso' },
-      { ripetizioni: '8', distanza: 100, recupero: 30, stile: 'rana' }
-    ];
 
     component.ngOnInit();
-
-    const req = httpMock.expectOne('/allenamenti');
-    expect(req.request.method).toEqual('GET');
-    req.flush(mockEsercizi);
-
-    expect(component.esercizi).toEqual(mockEsercizi);
   });
 });
