+++ vitoswimpt.client/src/app/esercizi/esercizi.component.html
@@ -1,1 +1,24 @@
-<p>esercizi works!</p>
+<h1 id="tableLabel">Esercizi Nuoto</h1>
+
+<p>Elenco Esercizi Registrati</p>
+
+<p *ngIf="!esercizi"><em>In caricamento... Ricarica la pagina dopo che il backend ASP.NET  è partito. Vedere <a href="https://aka.ms/jspsintegrationangular">https://aka.ms/jspsintegrationangular</a> per ulteriori dettagli.</em></p>
+
+<table *ngIf="esercizi">
+    <thead>
+        <tr>
+            <th>Ripetizioni</th>
+            <th>Distanza</th>
+            <th>Pausa</th>
+            <th>Stile</th>
+        </tr>
+    </thead>
+    <tbody>
+        <tr *ngFor="let esercizio of esercizi">
+          <td>{{ esercizio.ripetizioni }}x</td>
+          <td>{{ esercizio.distanza }}</td>
+          <td>{{ esercizio.recupero }}</td>
+          <td>{{ esercizio.stile }}</td>
+        </tr>
+    </tbody>
+</table>
