+++ vitoswimpt.client/src/app/esercizi/esercizi.component.ts
@@ -1,4 +1,12 @@
-import { Component } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+import { Component, OnInit } from '@angular/core';
+
+interface Esercizi {
+  ripetizioni: string;
+  distanza: number;
+  recupero: number;
+  stile: string;
+}
 
 @Component({
   selector: 'app-esercizi',
@@ -6,6 +14,26 @@ import { Component } from '@angular/core';
   templateUrl: './esercizi.component.html',
   styleUrl: './esercizi.component.css'
 })
-export class EserciziComponent {
+export class EserciziComponent implements OnInit {
 
+  public esercizi: Esercizi[] = [];
+
+  constructor(private http: HttpClient) { }
+
+  ngOnInit(): void {
+    this.getAllenamenti();
+  }
+
+  getAllenamenti() {
+    this.http.get<Esercizi[]>('/esercizi').subscribe(
+      (result) => {
+        this.esercizi = result;
+      },
+      (error) => {
+        console.error(error);
+      }
+    );
+  }
+
+  title = 'vitoswimpt.client';
 }
